# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API for managing artists, albums, and songs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Retrieve all songs",
                "description": "Returns a list of all songs",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "responses": {
                    "200": {
                        "description": "List of songs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Song"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Create a new song",
                "description": "Creates a new song and returns the created song object",
                "operationId": "2f1970b82d91cc715578d972189f19cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "album_id",
                                    "track_number"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Name of the song",
                                        "type": "string",
                                        "example": "Imagine"
                                    },
                                    "album_id": {
                                        "description": "Identifier of album this song belongs to",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "track_number": {
                                        "description": "Song index number in album",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Song successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Retrieve a song by ID",
                "description": "Fetches a song by their unique identifier",
                "operationId": "4325a96361d1d3820cc0580eaae9197c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the song to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Update a song",
                "description": "Updates an existing song by ID",
                "operationId": "452de7f8f11a121cd11e9430565632c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the song to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "album_id",
                                    "track_number"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Name of the song",
                                        "type": "string",
                                        "example": "Believe"
                                    },
                                    "album_id": {
                                        "description": "Identifier of album this song belongs to",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "track_number": {
                                        "description": "Song index number in album",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Song successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Song"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Delete a song",
                "description": "Deletes a song by ID",
                "operationId": "829c1a52ae5013ce3ff4b9829a31485f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the song to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Song successfully deleted"
                    },
                    "404": {
                        "description": "Song not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Song": {
                "title": "Song",
                "description": "Schema for Song",
                "required": [
                    "id",
                    "name",
                    "album_id",
                    "track_number"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the song",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the song",
                        "type": "string",
                        "example": "Imagine"
                    },
                    "album_id": {
                        "description": "Identifier of album this song belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "track_number": {
                        "description": "Song index number in album",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date when the song was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Date when the song was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}